Move SignedJWT extension functions to a separate file
